Collapse patterns over strings.
