+++ internal/goroutine/pool.go
@@ -53,3 +53,46 @@ func RunWorkersN(n int, worker PoolWorker) (err error) {
 
 	return err
 }
+
+// RunWorkersOverStrings invokes the given worker once for each of the
+// given string values. The worker function will receive the index as well
+// as the string value as parameters. Workers will be invoked in a number
+// of concurrent routines proportional to the maximum number of CPUs that
+// can be executing simultaneously.
+func RunWorkersOverStrings(values []string, worker func(index int, value string) error) error {
+	return RunWorkersOverStringsN(runtime.GOMAXPROCS(0), values, worker)
+}
+
+// RunWorkersOverStrings invokes the given worker once for each of the
+// given string values. The worker function will receive the index as well
+// as the string value as parameters. Workers will be invoked in n concurrent
+// routines.
+func RunWorkersOverStringsN(n int, values []string, worker func(index int, value string) error) error {
+	return RunWorkersN(n, indexedStringWorker(loadIndexedStringChannel(values), worker))
+}
+
+type indexedString struct {
+	Index int
+	Value string
+}
+
+func loadIndexedStringChannel(values []string) <-chan indexedString {
+	ch := make(chan indexedString, len(values))
+	defer close(ch)
+
+	for i, value := range values {
+		ch <- indexedString{Index: i, Value: value}
+	}
+
+	return ch
+}
+
+func indexedStringWorker(ch <-chan indexedString, worker func(index int, value string) error) PoolWorker {
+	return func(errs chan<- error) {
+		for value := range ch {
+			if err := worker(value.Index, value.Value); err != nil {
+				errs <- err
+			}
+		}
+	}
+}
